requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
google-generativeai
python-dotenv
streamlit
requests
pydantic
